@@ -1,13 +1,18 @@
 module github.com/teonet-go/teomq
 
-go 1.22.4
+go 1.24.0
 
-require github.com/teonet-go/teonet v0.6.6
+replace github.com/kirill-scherba/command/v2 => ../../kirill-scherba/command/v2
 
 require (
-	github.com/google/uuid v1.6.0 // indirect
+	github.com/kirill-scherba/command/v2 v2.3.0
+	github.com/teonet-go/teonet v0.6.6
+)
+
+require (
+	github.com/google/uuid v1.4.0 // indirect
 	github.com/kirill-scherba/bslice v0.0.2 // indirect
 	github.com/kirill-scherba/stable v0.0.8 // indirect
 	github.com/teonet-go/tru v0.0.18 // indirect
-	golang.org/x/sys v0.21.0 // indirect
+	golang.org/x/sys v0.14.0 // indirect
 )
