@@ -7,7 +7,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/teonet-go/teomq/commands"
+	"github.com/kirill-scherba/command/v2"
 	"github.com/teonet-go/teomq/consumer"
 	"github.com/teonet-go/teonet"
 )
@@ -73,20 +73,25 @@ func main() {
 }
 
 // Commands adds available broker commands.
-func Commands(c *commands.Commands) {
+func Commands(c *command.Commands) {
 	fmt.Println("Commands loaded:")
 
-	c.Add("version", "Get consumer version.", commands.Teonet, "{data}/{description}",
-		func(cmd *commands.CommandData, processIn commands.ProcessIn, data any) (
+	c.Add("version", "Get consumer version.", command.Teonet, "{data}/{description}",
+		"", "", "",
+		func(cmd *command.CommandData, processIn command.ProcessIn, data any) (
 			[]byte, error) {
 
 			// Parse teonet parameters
-			_, _, vars, err := c.Unmarshal(data.([]byte))
+			_, _, vars, d, err := c.ParseCommand(data.([]byte))
 			if err != nil {
 				return nil, err
 			}
 
-			return []byte(fmt.Sprintf("version: %s, data: %s", appVersion, vars["data"])), nil
+			log.Printf("process command %s, vars: %v, data_len: %d\n",
+				cmd.Cmd, vars, len(d))
+
+			return []byte(fmt.Sprintf("version: %s, data: %s",
+				appVersion, vars["data"])), nil
 		})
 
 	c.Print()
