@@ -7,8 +7,8 @@ import (
 	"io"
 	"log"
 
+	"github.com/kirill-scherba/command/v2"
 	"github.com/teonet-go/teomq/broker"
-	"github.com/teonet-go/teomq/commands"
 	"github.com/teonet-go/teonet"
 )
 
@@ -63,11 +63,11 @@ func main() {
 }
 
 // Commands adds available broker commands.
-func Commands(cmd *commands.Commands) {
+func Commands(cmd *command.Commands) {
 	fmt.Println("Commands loaded:")
 
-	cmd.Add("num_players", "Number of players.", commands.Teonet, "")
-	cmd.Add("num_servers", "Number of players.", commands.Teonet, "")
+	cmd.Add("num_players", "Number of players.", command.Teonet, "", "", "", "", nil)
+	cmd.Add("num_servers", "Number of players.", command.Teonet, "", "", "", "", nil)
 
 	cmd.Print()
 }
